Fix broken queue implementation
Whenever a second level cache entity is retrieved via the entity manager, the entity and the relations of the entity are stored in the entity manager. When the entity manager is cleared it means the entity is now unmanaged. When we're trying to get another entity via the unmanaged entity by a relation it breaks.

An example:
Fetch service (service and the relations are now managed)
Clear entity manager (service and relations are unmanaged)
Call service->getProfessions()

Doctrine tries to get the professions from second level cache. The owner of the cache item can't be found because the entity manager is cleared. That's why Doctrine breaks.